Update homework lesson26. Rename package name repository to services, class OrderRepository to OrderService and ProductRepository to ProductService.
