+++ src/main/java/edu/hillel/homework/lesson26/dao/ProductDAO.java
@@ -1,8 +1,8 @@
 package edu.hillel.homework.lesson26.dao;
 
-import edu.hillel.homework.lesson26.repository.ProductRepository;
+import edu.hillel.homework.lesson26.services.ProductService;
 
 public class ProductDAO {
 
-    public static final ProductRepository productRepository = new ProductRepository();
+    public static final ProductService productRepository = new ProductService();
 }
